derep: shift an unsigned operand for the abundance bound
clang-tidy's bugprone-signed-bitwise flagged the left shift in
resolvable_abundance, and it was the only instance of that warning in the
file, so 72a6a381 introduced it. 2^40 is well within int64_t, so nothing
was wrong at runtime; the fix is to shift the unsigned operand and convert
once, which is how the rest of the tree spells a shift feeding a wider type
(udb.cpp, random.cpp, os_byteswap.cpp all shift a uint*_t).

The type stays int64_t because the value is compared against s1, so a
uint64_t constant here would only trade this warning for a sign-compare one.

+++ src/core/derep.cpp
@@ -964,7 +964,7 @@ static auto dereplicating(std::unique_ptr<fastx_s> const & input_handle,
                       {
                         static constexpr auto uninformative_quality = int64_t{2};
                         static constexpr auto negligible_weight_ratio = int64_t{4};
-                        static constexpr auto resolvable_abundance = int64_t{1} << 40U;
+                        static constexpr auto resolvable_abundance = static_cast<int64_t>(uint64_t{1} << 40U);
                         auto const quality1 =
                           static_cast<int64_t>(symbol1) - parameters.opt_fastq_ascii;
                         if ((quality1 >= uninformative_quality)
